+++ Runtime/Api/Entitlement.cs
@@ -60,13 +60,15 @@ namespace AccelByte.Api
         /// </param>
         /// <param name="entitlementClazz">Class of the entitlement (optional)</param>
         /// <param name="entitlementAppType">This is the type of application that entitled (optional)</param>
-        public void QueryUserEntitlements( string entitlementName
+        /// <param name="features">The feature array</param>
+        public void QueryUserEntitlements(string entitlementName
             , string itemId
             , int offset
             , int limit
             , ResultCallback<EntitlementPagingSlicedResult> callback
             , EntitlementClazz entitlementClazz = EntitlementClazz.NONE
-            , EntitlementAppType entitlementAppType = EntitlementAppType.NONE )
+            , EntitlementAppType entitlementAppType = EntitlementAppType.NONE
+            , string[] features = null)
         {
             Report.GetFunctionLog(GetType().Name);
             Assert.IsNotNull(entitlementName, "Can't query user entitlements! EntitlementName parameter is null!");
@@ -83,6 +85,7 @@ namespace AccelByte.Api
                     session.UserId,
                     entitlementName,
                     itemId,
+                    features, 
                     offset,
                     limit,
                     entitlementClazz,