+++ Runtime/Server/ServerStatisticApi.cs
@@ -18,16 +18,16 @@ namespace AccelByte.Server
         /// <param name="httpClient"></param>
         /// <param name="config">baseUrl==StatisticServerUrl</param>
         /// <param name="session"></param>
-        internal ServerStatisticApi( IHttpClient httpClient
+        internal ServerStatisticApi(IHttpClient httpClient
             , ServerConfig config
-            , ISession session ) 
-            : base( httpClient, config, config.StatisticServerUrl, session )
+            , ISession session)
+            : base(httpClient, config, config.StatisticServerUrl, session)
         {
         }
 
-        public IEnumerator CreateUserStatItems( string userId
+        public IEnumerator CreateUserStatItems(string userId
             , CreateStatItemRequest[] statItems
-            , ResultCallback<StatItemOperationResult[]> callback )
+            , ResultCallback<StatItemOperationResult[]> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
@@ -46,7 +46,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
@@ -54,10 +54,10 @@ namespace AccelByte.Server
             callback.Try(result);
         }
 
-        public IEnumerator GetUserStatItems( string userId
+        public IEnumerator GetUserStatItems(string userId
             , ICollection<string> statCodes
             , ICollection<string> tags
-            , ResultCallback<PagedStatItems> callback )
+            , ResultCallback<PagedStatItems> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
@@ -85,7 +85,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<PagedStatItems>();
@@ -93,9 +93,9 @@ namespace AccelByte.Server
             callback.Try(result);
         }
 
-        public IEnumerator IncrementUserStatItems( string userId
+        public IEnumerator IncrementUserStatItems(string userId
             , StatItemIncrement[] data
-            , ResultCallback<StatItemOperationResult[]> callback )
+            , ResultCallback<StatItemOperationResult[]> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
@@ -114,7 +114,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
@@ -122,8 +122,8 @@ namespace AccelByte.Server
             callback.Try(result);
         }
 
-        public IEnumerator IncrementManyUsersStatItems( UserStatItemIncrement[] data
-            , ResultCallback<StatItemOperationResult[]> callback )
+        public IEnumerator IncrementManyUsersStatItems(UserStatItemIncrement[] data
+            , ResultCallback<StatItemOperationResult[]> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(data, nameof(data) + " cannot be null");
@@ -140,7 +140,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
@@ -148,9 +148,9 @@ namespace AccelByte.Server
             callback.Try(result);
         }
 
-        public IEnumerator ResetUserStatItems( string userId
+        public IEnumerator ResetUserStatItems(string userId
             , StatItemReset[] data
-            , ResultCallback<StatItemOperationResult[]> callback )
+            , ResultCallback<StatItemOperationResult[]> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
@@ -169,7 +169,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
@@ -177,8 +177,8 @@ namespace AccelByte.Server
             callback.Try(result);
         }
 
-        public IEnumerator ResetManyUsersStatItems( UserStatItemReset[] data
-            , ResultCallback<StatItemOperationResult[]> callback )
+        public IEnumerator ResetManyUsersStatItems(UserStatItemReset[] data
+            , ResultCallback<StatItemOperationResult[]> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(data, nameof(data) + " cannot be null");
@@ -195,7 +195,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
@@ -203,10 +203,10 @@ namespace AccelByte.Server
             callback.Try(result);
         }
 
-        public IEnumerator UpdateUserStatItems( string userId
+        public IEnumerator UpdateUserStatItems(string userId
             , string additionalKey
             , StatItemUpdate[] data
-            , ResultCallback<StatItemOperationResult[]> callback )
+            , ResultCallback<StatItemOperationResult[]> callback)
         {
             Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
             Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
@@ -226,7 +226,7 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
@@ -252,12 +252,199 @@ namespace AccelByte.Server
 
             IHttpResponse response = null;
 
-            yield return HttpClient.SendRequest(request, 
+            yield return HttpClient.SendRequest(request,
                 rsp => response = rsp);
 
             var result = response.TryParseJson<StatItemOperationResult[]>();
 
             callback.Try(result);
         }
-    }
+
+        public IEnumerator BulkFetchUserStatItemValues(string statCode
+           , string[] userIds
+           , string additionalKey
+           , ResultCallback<FetchUser[]> callback)
+        {
+            Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
+            Assert.IsNotNull(statCode, nameof(statCode) + " cannot be null");
+            Assert.IsNotNull(userIds, nameof(userIds) + " cannot be null");
+            Assert.IsNotNull(additionalKey, nameof(additionalKey) + " cannot be null");
+            Assert.IsNotNull(AuthToken, nameof(AuthToken) + " cannot be null");
+
+            var builder = HttpRequestBuilder
+                .CreateGet(BaseUrl + "/v2/admin/namespaces/{namespace}/statitems/value/bulk/getOrDefault")
+                .WithPathParam("namespace", Namespace_)
+                .WithBearerAuth(AuthToken)
+                .WithQueryParam("statCode", statCode)
+                .WithQueryParam("additionalKey", additionalKey)
+                .WithContentType(MediaType.ApplicationJson)
+                .Accepts(MediaType.ApplicationJson);
+            foreach (var userId in userIds)
+            {
+                builder.WithQueryParam("userIds", userId);
+            }
+            var request = builder.GetResult();
+
+            IHttpResponse response = null;
+
+            yield return HttpClient.SendRequest(request,
+                rsp => response = rsp);
+
+            var result = response.TryParseJson<FetchUser[]>();
+
+            callback.Try(result);
+        }
+
+        public IEnumerator BulkUpdateMultipleUserStatItemsValue(UpdateUserStatItem[] bulkUpdateMultipleUserStatItem
+           , ResultCallback<UpdateUserStatItemsResponse[]> callback)
+        {
+            Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
+            Assert.IsNotNull(bulkUpdateMultipleUserStatItem, nameof(bulkUpdateMultipleUserStatItem) + " cannot be null");
+            Assert.IsNotNull(AuthToken, nameof(AuthToken) + " cannot be null");
+
+            var request = HttpRequestBuilder
+            .CreatePut(BaseUrl + "/v2/admin/namespaces/{namespace}/statitems/value/bulk")
+            .WithPathParam("namespace", Namespace_)
+            .WithBearerAuth(AuthToken)
+            .WithContentType(MediaType.ApplicationJson)
+            .WithBody(bulkUpdateMultipleUserStatItem.ToUtf8Json())
+            .Accepts(MediaType.ApplicationJson)
+            .GetResult();
+
+            IHttpResponse response = null;
+
+            yield return HttpClient.SendRequest(request,
+                rsp => response = rsp);
+
+            var result = response.TryParseJson<UpdateUserStatItemsResponse[]>();
+
+            callback.Try(result);
+        }
+
+        public IEnumerator BulkResetUserStatItemsValues(string userId
+            , string additionalKey
+            , UserStatItem[] bulkUserStatItems
+           , ResultCallback<UpdateUserStatItemsResponse[]> callback)
+        {
+            Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
+            Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
+            Assert.IsNotNull(additionalKey, nameof(additionalKey) + " cannot be null");
+            Assert.IsNotNull(bulkUserStatItems, nameof(bulkUserStatItems) + " cannot be null");
+            Assert.IsNotNull(AuthToken, nameof(AuthToken) + " cannot be null");
+
+            var request = HttpRequestBuilder
+            .CreatePut(BaseUrl + "/v2/admin/namespaces/{namespace}/users/{userId}/statitems/value/reset/bulk")
+            .WithPathParam("namespace", Namespace_)
+            .WithBearerAuth(AuthToken)
+            .WithPathParam("userId", userId)
+            .WithQueryParam("additionalKey", additionalKey)
+            .WithContentType(MediaType.ApplicationJson)
+            .WithBody(bulkUserStatItems.ToUtf8Json())
+            .Accepts(MediaType.ApplicationJson)
+            .GetResult();
+
+            IHttpResponse response = null;
+
+            yield return HttpClient.SendRequest(request,
+                rsp => response = rsp);
+
+            var result = response.TryParseJson<UpdateUserStatItemsResponse[]>();
+
+            callback.Try(result);
+        }
+
+        public IEnumerator BulkUpdateUserStatItemValue(string userId
+           , string additionalKey
+           , UpdateUserStatItemWithStatCode[] BulkUpdateUserStatItem
+           , ResultCallback<UpdateUserStatItemsResponse[]> callback)
+        {
+            Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
+            Assert.IsNotNull(userId, nameof(userId) + " cannot be null");
+            Assert.IsNotNull(additionalKey, nameof(additionalKey) + " cannot be null");
+            Assert.IsNotNull(BulkUpdateUserStatItem, nameof(BulkUpdateUserStatItem) + " cannot be null");
+            Assert.IsNotNull(AuthToken, nameof(AuthToken) + " cannot be null");
+
+            var request = HttpRequestBuilder
+            .CreatePut(BaseUrl + "/v2/admin/namespaces/{namespace}/users/{userId}/statitems/value/bulk")
+            .WithPathParam("namespace", Namespace_)
+            .WithBearerAuth(AuthToken)
+            .WithPathParam("userId", userId)
+            .WithQueryParam("additionalKey", additionalKey)
+            .WithContentType(MediaType.ApplicationJson)
+            .WithBody(BulkUpdateUserStatItem.ToUtf8Json())
+            .Accepts(MediaType.ApplicationJson)
+            .GetResult();
+
+            IHttpResponse response = null;
+
+            yield return HttpClient.SendRequest(request,
+                rsp => response = rsp);
+
+            var result = response.TryParseJson<UpdateUserStatItemsResponse[]>();
+
+            callback.Try(result);
+        }
+
+
+        public IEnumerator UpdateUserStatItemValue(string userId, string statCode
+           , string additionalKey
+           , UpdateUserStatItem UpdateUserStatItemValue
+           , ResultCallback<UpdateUserStatItemValueResponse> callback)
+        {
+            Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
+            Assert.IsNotNull(statCode, nameof(statCode) + " cannot be null");
+            Assert.IsNotNull(additionalKey, nameof(additionalKey) + " cannot be null");
+            Assert.IsNotNull(AuthToken, nameof(AuthToken) + " cannot be null");
+
+            var request = HttpRequestBuilder
+            .CreatePut(BaseUrl + "/v2/admin/namespaces/{namespace}/users/{userId}/stats/{statCode}/statitems/value")
+            .WithPathParam("namespace", Namespace_)
+            .WithBearerAuth(AuthToken)
+            .WithPathParam("userId", userId)
+            .WithPathParam("statCode", statCode)
+            .WithQueryParam("additionalKey", additionalKey)
+            .WithBody(UpdateUserStatItemValue.ToUtf8Json())
+            .WithContentType(MediaType.ApplicationJson)
+            .Accepts(MediaType.ApplicationJson)
+            .GetResult();
+
+            IHttpResponse response = null;
+
+            yield return HttpClient.SendRequest(request,
+                rsp => response = rsp);
+
+            var result = response.TryParseJson<UpdateUserStatItemValueResponse>();
+
+            callback.Try(result);
+        }
+
+        public IEnumerator DeleteUserStatItems(string userId, string statCode
+           , string additionalKey
+           , ResultCallback callback)
+        {
+            Assert.IsNotNull(Namespace_, nameof(Namespace_) + " cannot be null");
+            Assert.IsNotNull(statCode, nameof(statCode) + " cannot be null");
+            Assert.IsNotNull(additionalKey, nameof(additionalKey) + " cannot be null");
+            Assert.IsNotNull(AuthToken, nameof(AuthToken) + " cannot be null");
+
+            var request = HttpRequestBuilder
+            .CreateDelete(BaseUrl + "/v2/admin/namespaces/{namespace}/users/{userId}/stats/{statCode}/statitems")
+            .WithPathParam("namespace", Namespace_)
+            .WithBearerAuth(AuthToken)
+            .WithPathParam("userId", userId)
+            .WithPathParam("statCode", statCode)
+            .WithQueryParam("additionalKey", additionalKey)
+            .WithContentType(MediaType.ApplicationJson)
+            .Accepts(MediaType.ApplicationJson)
+            .GetResult();
+
+            IHttpResponse response = null;
+
+            yield return HttpClient.SendRequest(request,
+                rsp => response = rsp);
+
+            var result = response.TryParse();
+            callback.Try(result);
+        }
+    }        
 }